Added main vision code (so far)

+++ src/com/_2491nomythic/util/Scores.java
@@ -1,7 +1,7 @@
 package com._2491nomythic.util;
 
 public class Scores {
-	public int rectangularity;
-	public int aspectRatioVertical;
-	public int aspectRatioHorizontal;
+	public double rectangularity;
+	public double aspectRatioVertical;
+	public double aspectRatioHorizontal;
 }
